Lisa:
Normally, a content creator is in the best position to know how strong the protective signing ought to be on their content. Thus, a content creator can choose weak or strong cryptography, and a content consumer would normally accept that. There MAY be applications where the content consumer chooses to reject weak cryptography, but that is not envisioned as the common use case. 

git-svn-id: svn://svn.code.sf.net/p/metalinks/code@574 84cd5247-9d1f-0410-ad08-ff3d2110de5e